items crossing off list
