+++ src/App.js
@@ -1,9 +1,12 @@
 import React from 'react';
-import Todo from './Todo';
+import TodoList from './components/TodoList';
+import Todo from './components/Todo';
+
 
 const App = () => (
   <div>
     <h2>Todo App</h2>
+    <TodoList />
     <Todo />
   </div>
 );
