+++ src/components/Todo.js
@@ -1,0 +1,22 @@
+import React, { Component } from "react";
+
+class Todo extends Component {
+    constructor() {
+        super();
+
+        this.state = {
+            clicked: false,
+        }
+    }
+    handleClick = () => {
+        this.setState({ clicked: !this.state.clicked });
+    }
+    render() {
+        const styles = this.state.clicked ? { textDecoration: 'line-through' } : { textDecoration: 'none' };
+        return (
+            <div style={styles} onClick={this.handleClick}>{this.props.items}</div>
+        );
+    }
+}
+
+export default Todo;
